net, netpod: Move netpod related source files to a new package
Relocate the netpod source files into a new sub-package.

netpod is currently handling only the configuration stage.
The addition of the discovery stage will increase the amount of code and
therefore requires a small re-arrangement to accommodate the new logic.
