+++ pkg/network/setup/BUILD.bazel
@@ -6,7 +6,6 @@ go_library(
         "configstate.go",
         "configstatecache.go",
         "netconf.go",
-        "netpod.go",
         "netstat.go",
         "network.go",
         "podnic.go",
@@ -19,19 +18,16 @@ go_library(
         "//pkg/network/dhcp:go_default_library",
         "//pkg/network/domainspec:go_default_library",
         "//pkg/network/driver:go_default_library",
-        "//pkg/network/driver/nmstate:go_default_library",
-        "//pkg/network/driver/procsys:go_default_library",
         "//pkg/network/errors:go_default_library",
         "//pkg/network/infraconfigurators:go_default_library",
         "//pkg/network/istio:go_default_library",
         "//pkg/network/link:go_default_library",
         "//pkg/network/namescheme:go_default_library",
-        "//pkg/network/netmachinery:go_default_library",
         "//pkg/network/netns:go_default_library",
-        "//pkg/network/setup/masquerade:go_default_library",
+        "//pkg/network/setup/netpod:go_default_library",
+        "//pkg/network/setup/netpod/masquerade:go_default_library",
         "//pkg/network/sriov:go_default_library",
         "//pkg/network/vmispec:go_default_library",
-        "//pkg/pointer:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
         "//pkg/virt-launcher/virtwrap/converter:go_default_library",
@@ -51,7 +47,6 @@ go_test(
         "configstate_test.go",
         "configstatecache_test.go",
         "netconf_test.go",
-        "netpod_test.go",
         "netstat_test.go",
         "network_suite_test.go",
         "network_test.go",
@@ -64,15 +59,12 @@ go_test(
         "//pkg/network/cache:go_default_library",
         "//pkg/network/dhcp:go_default_library",
         "//pkg/network/driver:go_default_library",
-        "//pkg/network/driver/nmstate:go_default_library",
-        "//pkg/network/driver/procsys:go_default_library",
         "//pkg/network/errors:go_default_library",
         "//pkg/network/infraconfigurators:go_default_library",
         "//pkg/network/namescheme:go_default_library",
         "//pkg/network/sriov:go_default_library",
         "//pkg/network/vmispec:go_default_library",
         "//pkg/os/fs:go_default_library",
-        "//pkg/pointer:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/api:go_default_library",
