+++ pkg/network/setup/netpod/BUILD.bazel
@@ -2,29 +2,34 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["masquerade.go"],
-    importpath = "kubevirt.io/kubevirt/pkg/network/setup/masquerade",
+    srcs = ["netpod.go"],
+    importpath = "kubevirt.io/kubevirt/pkg/network/setup/netpod",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/network/driver/nft:go_default_library",
         "//pkg/network/driver/nmstate:go_default_library",
-        "//pkg/network/istio:go_default_library",
+        "//pkg/network/driver/procsys:go_default_library",
+        "//pkg/network/link:go_default_library",
+        "//pkg/network/namescheme:go_default_library",
         "//pkg/network/netmachinery:go_default_library",
-        "//pkg/util/net/ip:go_default_library",
+        "//pkg/network/setup/netpod/masquerade:go_default_library",
+        "//pkg/network/vmispec:go_default_library",
+        "//pkg/pointer:go_default_library",
+        "//pkg/virt-launcher/virtwrap/api:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
+        "//staging/src/kubevirt.io/client-go/log:go_default_library",
     ],
 )
 
 go_test(
     name = "go_default_test",
     srcs = [
-        "masquerade_suite_test.go",
-        "masquerade_test.go",
+        "netpod_suite_test.go",
+        "netpod_test.go",
     ],
     deps = [
         ":go_default_library",
-        "//pkg/network/driver/nft:go_default_library",
         "//pkg/network/driver/nmstate:go_default_library",
+        "//pkg/network/driver/procsys:go_default_library",
         "//pkg/pointer:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/testutils:go_default_library",
