+++ pkg/network/setup/netpod/masquerade/BUILD.bazel
@@ -1,0 +1,34 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["masquerade.go"],
+    importpath = "kubevirt.io/kubevirt/pkg/network/setup/netpod/masquerade",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/network/driver/nft:go_default_library",
+        "//pkg/network/driver/nmstate:go_default_library",
+        "//pkg/network/istio:go_default_library",
+        "//pkg/network/netmachinery:go_default_library",
+        "//pkg/util/net/ip:go_default_library",
+        "//staging/src/kubevirt.io/api/core/v1:go_default_library",
+    ],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = [
+        "masquerade_suite_test.go",
+        "masquerade_test.go",
+    ],
+    deps = [
+        ":go_default_library",
+        "//pkg/network/driver/nft:go_default_library",
+        "//pkg/network/driver/nmstate:go_default_library",
+        "//pkg/pointer:go_default_library",
+        "//staging/src/kubevirt.io/api/core/v1:go_default_library",
+        "//staging/src/kubevirt.io/client-go/testutils:go_default_library",
+        "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
+        "//vendor/github.com/onsi/gomega:go_default_library",
+    ],
+)
